+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog
 0.62 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added `setup.load_type_from_package` to reload a single type.
+  Moved `workflow.load_workflow_from_package` to `setup.load_workflow_from_package`.
+  [gbastien]
 
 0.61 (2022-07-01)
 -----------------